backend/requirements.txt
fastapi==0.115.0
uvicorn==0.30.6
python-jose==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==3.2.2
pydantic==2.9.2
pydantic-settings==2.6.0
email-validator==2.2.0
SQLAlchemy==2.0.36
alembic==1.13.2
psycopg2-binary==2.9.9
scikit-learn==1.5.2
pandas==2.2.3
numpy==2.1.2
joblib==1.4.2
websockets==12.0
httpx==0.27.2
pytest==8.3.3
